+++ src/settings.h
@@ -68,7 +68,7 @@ public:
 	int getWindowWidth() {return m_windowWidth;};
 	int getWindowHeight() {return m_windowHeight;};
   //grid
-	int getGridSize() {return m_gridSize;};
+	size_t getGridSize() {return m_gridSize;};
 	int getGridWidth() {return m_windowWidth / TILE_HALF_WIDTH;};
 	int getGridHeight() {return m_windowHeight / TILE_HALF_HEIGHT;};
 	int getGridNbrOfLine() {return m_gridNbrOfLine;};
@@ -106,7 +106,7 @@ private:
 	int m_windowWidth;
 	int m_windowHeight;
   //grid
-	int m_gridSize; //the grid is a square even if the screen is non-square.
+	size_t m_gridSize; //the grid is a square even if the screen is non-square.
 	int m_gridNbrOfLine; //nbr of tiles in a line
 	int m_gridNbrOfRow; //nbr of tiles in a row
 	int m_first_tile_x; //first tile to display
